+++ src/backend/marsha/core/tests/test_api_get_frontend_configuration.py
@@ -11,6 +11,9 @@ from marsha.core.factories import SiteConfigFactory
     SENTRY_DSN="https://sentry.dsn",
     ENVIRONMENT="development",
     RELEASE="1.2.3",
+    P2P_ENABLED=True,
+    P2P_STUN_SERVER_URLS=["stun:stun.l.google.com:19302"],
+    P2P_WEB_TORRENT_TRACKER_URLS=["wss://tracker.webtorrent.io"],
 )
 class TestGetFrontendConfiguration(TestCase):
     """Test the get_frontend_configuration API."""
@@ -34,6 +37,11 @@ class TestGetFrontendConfiguration(TestCase):
                 "release": "1.2.3",
                 "sentry_dsn": "https://sentry.dsn",
                 "inactive_resources": [],
+                "p2p": {
+                    "isEnabled": True,
+                    "stunServerUrls": ["stun:stun.l.google.com:19302"],
+                    "webTorrentTrackerUrls": ["wss://tracker.webtorrent.io"],
+                },
             },
         )
 
@@ -54,6 +62,11 @@ class TestGetFrontendConfiguration(TestCase):
                 "release": "1.2.3",
                 "sentry_dsn": None,
                 "inactive_resources": [],
+                "p2p": {
+                    "isEnabled": True,
+                    "stunServerUrls": ["stun:stun.l.google.com:19302"],
+                    "webTorrentTrackerUrls": ["wss://tracker.webtorrent.io"],
+                },
             },
         )
 
@@ -73,6 +86,11 @@ class TestGetFrontendConfiguration(TestCase):
                 "release": "1.2.3",
                 "sentry_dsn": "https://sentry.dsn",
                 "inactive_resources": [],
+                "p2p": {
+                    "isEnabled": True,
+                    "stunServerUrls": ["stun:stun.l.google.com:19302"],
+                    "webTorrentTrackerUrls": ["wss://tracker.webtorrent.io"],
+                },
             },
         )
 
@@ -93,5 +111,10 @@ class TestGetFrontendConfiguration(TestCase):
                 "release": "1.2.3",
                 "sentry_dsn": "https://sentry.dsn",
                 "inactive_resources": ["video"],
+                "p2p": {
+                    "isEnabled": True,
+                    "stunServerUrls": ["stun:stun.l.google.com:19302"],
+                    "webTorrentTrackerUrls": ["wss://tracker.webtorrent.io"],
+                },
             },
         )