+++ src/backend/marsha/core/views.py
@@ -126,6 +126,11 @@ class MarshaViewMixin:
             },
             "release": settings.RELEASE,
             "sentry_dsn": settings.SENTRY_DSN,
+            "p2p": {
+                "isEnabled": settings.P2P_ENABLED,
+                "stunServerUrls": settings.P2P_STUN_SERVER_URLS,
+                "webTorrentTrackerUrls": settings.P2P_WEB_TORRENT_TRACKER_URLS,
+            },
             "static": {
                 "img": {
                     "liveBackground": static("img/liveBackground.jpg"),