+++ src/frontend/apps/lti_site/components/App/AppInitializer/index.spec.tsx
@@ -8,6 +8,7 @@ import {
   useThumbnail,
   useSharedLiveMedia,
   useDocument,
+  useP2PConfig,
 } from 'lib-components';
 import { useAttendance } from 'lib-video';
 import React from 'react';
@@ -32,6 +33,11 @@ jest.mock('lib-components', () => ({
     video: mockedVideo,
     document: { id: 'my-document-id' },
     attendanceDelay: 6,
+    p2p: {
+      isEnabled: true,
+      stunServerUrls: ['stun.example.com'],
+      webTorrentTrackerUrls: ['tracker.example.com'],
+    },
   }),
   decodeJwt: () => ({
     maintenance: true,
@@ -44,6 +50,9 @@ describe('<AppInitializer />', () => {
       setSentry: () => useSentry.setState({ isSentryReady: true }),
     });
 
+    expect(useP2PConfig.getState().isP2PEnabled).toEqual(false);
+    expect(useP2PConfig.getState().stunServersUrls).toEqual([]);
+    expect(useP2PConfig.getState().webTorrentServerTrackerUrls).toEqual([]);
     expect(useSentry.getState().isSentryReady).toEqual(false);
     expect(useVideo.getState().videos).toEqual({});
     expect(useTimedTextTrack.getState().timedtexttracks).toEqual({});
@@ -63,6 +72,11 @@ describe('<AppInitializer />', () => {
 
     await screen.findByText('some cool content');
 
+    expect(useP2PConfig.getState().isP2PEnabled).toEqual(true);
+    expect(useP2PConfig.getState().stunServersUrls).toEqual(['stun.example.com']);
+    expect(useP2PConfig.getState().webTorrentServerTrackerUrls).toEqual([
+      'tracker.example.com',
+    ])
     expect(useSentry.getState().isSentryReady).toEqual(true);
     expect(useVideo.getState().videos).toEqual({
       ['my-video-id']: mockedVideo,