+++ src/frontend/packages/lib_components/src/data/stores/index.tsx
@@ -6,3 +6,4 @@ export * from './useSharedLiveMedia';
 export * from './useDocument';
 export * from './useSentry';
 export * from './generics';
+export * from './useP2PVideoConfig';